OH: Feat: Shopping 패키지 컨트롤러 레이어 커스텀 예외 관련 수정

+++ src/main/java/com/spring/shopping/service/WishlistServiceImpl.java
@@ -47,6 +47,9 @@ public class WishlistServiceImpl implements WishlistService {
 
     @Override
     public void removeFromWishlist(Long wishlistId) {
+        if(wishlistId == null) {
+            throw new CustomException(ExceptionCode.WISHLIST_ID_NOT_FOUND);
+        }
         wishlistRepository.deleteById(wishlistId);
     }
